[For AMO] Fix extension's preferences to be prefixed with 'extensions.refwdformatter.'
To meet the criteria for AMO ( addons.mozilla.org ) .
--- Reviewer comments:
`pref` called in potentially dangerous manner. Extensions should not
alter preferences outside of the 'extensions.' preference branch. Please
make sure that all of your extension's preferences are prefixed with
'extensions.add-on-name.', where 'add-on-name' is a distinct string
unique to and indicative of your add-on.
---